pyproject.toml
[tool.poetry]
name = "blog"
version = "0.1.0"
description = ""
authors = ["Dmitry Golub <gosest4@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standart"], version = "^0.115.11"}
pydantic = "^2.10.6"
pydantic-settings = "^2.8.1"
alembic = "^1.15.1"
asyncpg = "^0.30.0"
sqlalchemy = "^2.0.39"
python-jose = "^3.4.0"
passlib = "^1.7.4"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.2"}
celery = "^5.5.0"
flower = "^2.0.1"
pytest-asyncio = "^0.26.0"
httpx = "^0.28.1"
python-json-logger = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
